Extract utils from dev scripts
