+++ packages/chookscord/src/utils/register-commands.ts
